% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RkMetrics_functions.R
\name{lplot}
\alias{lplot}
\title{A Plotting Function}
\usage{
lplot(n, x, add = F)
}
\arguments{
\item{n}{the length of the vector Defaults to TRUE.}

\item{x}{the vector arguement.}

\item{add}{whether to add lines. Default is FALSE}
}
\description{
Produces a plot of the area-under-the-curve for the mortality data
}
\examples{
m1 <- Mortality$D.Male[which(Mortality$Year == 2008)]
m2 <- Mortality$E.Male[which(Mortality$Year == 2008)]
male.1 <- m1/m2
male.2 <- log(male.1[!is.na(male.1)])
lplot(1:length(male.2),male.2)

lplot(1:length(male.2),male.2,add=TRUE)

}
\keyword{lplot}
