% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_Doc.R
\docType{data}
\name{Mortality}
\alias{Mortality}
\title{Switzerland Mortality Data}
\format{A data frame with 6 columns corresponding to:
\describe{
  \item{Year}{Corresponding year of , in US dollars}
  \item{Age}{Age of the individual}
  \item{E.Male}{Male Exposed-to-Risk Population}
  \item{E.Female}{Female Exposed-to-Risk Population}
  \item{D.Male}{Number of male death counts, for the given year and age}
  \item{D.Female}{Number of female death counts, for the given year and age}
}}
\source{
\url{http://www.mortality.org/cgi-bin/hmd/country.php?cntr=CHE&level=1}
}
\usage{
Mortality
}
\description{
Exposed to Risk and number of deaths data.
}
\details{
Mortality data for both Males and Females in Switzerland, from 1981
to 2014.

These data are freely available at the Human Mortality Database
}
\references{
Glei, D. and Andreeva, M. (2016). About mortality data for switzerland.
}
\keyword{datasets}
