% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RkMetrics_functions.R
\name{pccopula}
\alias{pccopula}
\title{Produces a plot of a copula, which can be used to assess the dependency between
two sexes bounded by the actual and the expanded mortality estimates}
\usage{
pccopula(theta, pl = 1, z)
}
\arguments{
\item{theta}{gives the order.}

\item{pl}{gives the association.}

\item{z}{the length of the z axis Defaults to 10.}
}
\description{
Produces a plot of a copula, which can be used to assess the dependency between
two sexes bounded by the actual and the expanded mortality estimates
}
\examples{
#Examples

pccopula(theta=3,pl=.5,z=10)


}
\keyword{pccopula()}
