% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RkMetrics_functions.R
\name{pgcopula}
\alias{pgcopula}
\title{Similar to pccopula(), but suitable when the dependence is
stronger at the older ages}
\usage{
pgcopula(theta, pl = 1, z)
}
\arguments{
\item{theta}{gives the order.}

\item{pl}{gives the association, with a correction for the direction of dependence}

\item{z}{the length of the z axis Defaults to 10.}
}
\description{
Similar to pccopula(), but suitable when the dependence is
stronger at the older ages
}
\examples{
#Examples

pgcopula(theta=1.3,pl=2,z=10)


}
\keyword{pgcopula()}
