% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RkMetrics_functions.R
\name{vplot}
\alias{vplot}
\title{Produces a similar plot as lplot(), only a transposition of ages is made}
\usage{
vplot(n, x, add = F)
}
\arguments{
\item{n}{the length of the vector Defaults to TRUE.}

\item{x}{the vector arguement.}

\item{add}{whether to add lines. Default is FALSE}
}
\description{
Produces a similar plot as lplot(), only a transposition of ages is made
}
\examples{
m1 <- Mortality$D.Male[which(Mortality$Year == 2008)]
m2 <- Mortality$E.Male[which(Mortality$Year == 2008)]
male.1 <- m1/m2
male.2 <- log(male.1[!is.na(male.1)])
vplot(1:length(male.2),male.2)

}
\keyword{vplot}
